+++ ke-legacy/main.js
@@ -1036,16 +1036,27 @@ var PaymentMenuText = function (client){
     if (GetLang()){sayText('You are paying into account number '+client.AccountNumber+' Total Repaid '+repaid+', Bal '+balance+'.Please reply with the amount you want to pay');}
     else {sayText('Unafanya malipo kwa hii akaunti '+client.AccountNumber+' Jumla ya malipo '+repaid+', Salio '+balance+'.Tafadhali weka kiasi unachotaka kulipa');}
 };
-var CheckBalanceMenuText = function (Overpaid,Season,Credit,Paid,Balance){
+var CheckBalanceMenuText = function (healthyPathDetails, Overpaid,Season,Credit,Paid,Balance){
+    var getHealthyPathMessage = require('../healthy-path/balance/healthyPathOnBalance');
+    var balanceMenu = '';
+    var healthyPathMessage = '';
     if (GetLang()){
-        if(Overpaid){sayText(Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nOver payment: '+Balance+ '\n1 - Make payment');}
-        else {sayText(Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nRemaining: '+Balance+ '\n1 - Make payment');}
+        healthyPathMessage = getHealthyPathMessage(healthyPathDetails.seasonId, healthyPathDetails.countryId, healthyPathDetails.districtId, Credit, Paid, 'en');
+        if(Overpaid){
+            balanceMenu = Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nOver payment: '+Balance+ '\n' + healthyPathMessage + '1 - Make payment';
+        }else {
+            balanceMenu = Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nRemaining: '+Balance + '\n' + healthyPathMessage +  '1 - Make payment';
+        }
+    } else{
+        healthyPathMessage = getHealthyPathMessage(Credit, Paid, 'sw');
+        if(Overpaid){
+            balanceMenu = Season+':\nJumla ya malipo: '+Paid+'\nJumla ya mkopo: '+Credit+'\nMalipo kwa mkopo unaofuata: '+Balance+ '\n' + healthyPathMessage + '1 - Fanya malipo';
+        }else {
+            balanceMenu = Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nSalio: '+Balance+ '\n' + healthyPathMessage + '1 - Fanya malipo';
+        }
     }
-    else{
-        if(Overpaid){sayText(Season+':\nJumla ya malipo: '+Paid+'\nJumla ya mkopo: '+Credit+'\nMalipo kwa mkopo unaofuata: '+Balance+ '\n1 - Fanya malipo');}
-        else {sayText(Season+':\nPaid: '+Paid+'\nTotal credit: '+Credit+'\nSalio: '+Balance+ '\n1 - Fanya malipo');}
-    }
-    var BalanceInfo = 'Balance: '+Balance+ '\nSeason: '+Season+ '\nCredit: '+Credit+ '\nPaid: '+Paid+ '\nOverpaid: '+Overpaid;
+    sayText(balanceMenu);
+    var BalanceInfo = 'Balance: '+Balance+ '\nSeason: '+Season+ '\nCredit: '+Credit+ '\nPaid: '+Paid+ '\nOverpaid: '+Overpaid + '\n' + healthyPathMessage;
     call.vars.BalanceInfo = BalanceInfo;
     notifyELK();
 };
@@ -1874,7 +1885,15 @@ addInputHandler('MainMenu', function(SplashMenu){
                 }
             }
         }
-        CheckBalanceMenuText (Overpaid,Season,Credit,Paid,Balance);
+        var mostRecentSeason = client.BalanceHistory[0];
+        var healthyPathDetails = {
+            countryId: client.CountryId,
+            districtId: client.DistrictId
+        };
+        if (mostRecentSeason) {
+            healthyPathDetails.seasonId = mostRecentSeason.SeasonId;
+        }
+        CheckBalanceMenuText (healthyPathDetails, Overpaid,Season,Credit,Paid,Balance);
         promptDigits('ContinueToPayment', {submitOnHash: true, maxDigits: 1, timeout: 5});
     }
 });